requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-WTF==1.1.1
Pillow==10.0.1
Werkzeug==2.3.7
python-socketio==5.9.0
eventlet==0.33.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
cloudinary==1.33.0
python-dotenv==1.0.0
requests==2.31.0